Let click bot inherit from parent
