+++ clickbot/Dockerfile
@@ -1,11 +1,11 @@
 FROM python:3.9
 
-RUN mkdir -p /usr/src/clickbot
-WORKDIR /usr/src/clickbot
+RUN mkdir -p /usr/src
+WORKDIR /usr/src
 
-COPY clickbot/requirements.txt /usr/src/clickbot
-RUN pip install --no-cache-dir -r requirements.txt
-
+COPY templates.py /usr/src/
 COPY clickbot /usr/src/clickbot
 
-ENTRYPOINT ["python", "clickbot.py"]
+RUN pip install --no-cache-dir -r clickbot/requirements.txt
+
+ENTRYPOINT ["python", "-m", "clickbot"]
